+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 
 # x.x.x
 
+* Switch from scikit-image to scikit-learn for ransac
 * Add warning in minmax returns a NaN
 * Add unittests from theory for FFT, scheludko, minmax
 * FFT: code refactoring